Ditch intermediate kernel composition in favour of direct graph operations (#35)
This change implements a substantial change to the core mechanism by which Stheno determines the mean vector and covariance matrix of a GP. The previous mechanism was to derive a composite kernel from a GP, and evaluate it at a number of input locations to obtain the covariance of or between and pair of GPs. This new implementation favours a more direct approach, that broadly avoids this intermediate representation.

The documentation should be consulted for more information, but this change has significantly simplified the codebase with (for all intents and purposes) not loss of functionality. It also provides far more readable stack traces, and should improve compile times, just because there's far less stuff to compile than before.

* Initial pass. Product completed

* Implement + test product with new approach

* Conditioning and indexing

* Implement cross transformation

* Implement composition

* Improve sampling stability

* Fix up some FiniteGP tests

* Test product kernel

* Some work

* Fix cross functionality

* Remove redundant code

* Fix up conditional sampling crash

* Clean up some testing

* Revert sampling behaviour and fix Distances dims warnings

* Tests for conditional

* First stab at some documentation for the new set up

* Start change to new interface + test + documentation

* Add more cov_diag tests + fix conditioning bug

* Style

* composition consistency tests

* Also require that CI passes for Julia 1.2

* Minor coverage improvements

* Untested implementation of approximate condtioning

* Complete elbo / Titsias + improve coverage

* Tweaks docs + stretch kernel

* Docs + minor bug fix

* Move some stuff around

* Re-add convenience functionality

* Enable tests

* Documentation hosting + badge